Create pico_reader.py
This is a work-in-progress PicoDst reader for Python, along with some simple event cuts (track cuts are a work in progress as well). Right now it only reads in the few things I need for doing a proton kurtosis analysis, but it can (and should) be made more robust. Also, some of the definitions in it could be offloaded to other objects/macros (like in STAR's PicoDst reader), but I don't know how much time that would save. For now, it works for what it works on and not much else.